_ROS2_FreshmanProject/requirements.txt
pyserial
scipy
numpy
